requirements.txt
torch>=2.0.0
openai-whisper
pyannote.audio
pydub
numpy